Step 3 (#249)
* print 책임 리팩토링

* step-3 부분 구현 (예외처리 미구현)

+++ src/main/java/domain/LottoResults.java
@@ -19,7 +19,7 @@ public class LottoResults {
         lottoResults.put(lottoResult, lottoResults.get(lottoResult) + 1);
     }
 
-    public long getCountBy(LottoResult lottoResult) {
+    public int getCountBy(LottoResult lottoResult) {
         return lottoResults.getOrDefault(lottoResult, 0);
     }
 }
